update some responseUtils to ts
